backend/agents/regulatory_agent.py
"""
FinSight — SEBI Regulatory Radar Agent.
Monitors live regulatory filings (SEBI, BSE) and assesses materiality/risk using an LLM.
"""

from __future__ import annotations

import asyncio
import json
import logging
import os
from typing import Any, Optional
from datetime import datetime, timezone

import feedparser
import httpx

from backend.models.schemas import RegulatoryData

logger = logging.getLogger(__name__)

# ── Configuration ───────────────────────────────────────────────────
OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY", "")
OPENROUTER_BASE_URL = os.getenv("OPENROUTER_BASE_URL", "https://openrouter.ai/api/v1")
REGULATORY_MODEL = os.getenv("REGULATORY_MODEL", "google/gemini-2.0-flash")

_SYSTEM_PROMPT = (
    "You are a regulatory risk analyst for the Indian stock market. "
    "Analyze the following regulatory filings or corporate announcements. "
    "Return ONLY valid JSON with these exact keys: "
    "sentiment_impact (exactly one of: positive, negative, neutral), "
    "max_risk_score (float 0.0 to 10.0), "
    "categorized_events (array of objects with keys: category, risk_score, description), "
    "signal (exactly one of: BUY, SELL, HOLD), "
    "confidence (float 0.0 to 1.0), "
    "reasoning (string, max 2 sentences). "
    "Do not include any text outside the JSON object."
)

def _get_regulatory_feeds(symbol: str) -> list[str]:
    """
    Fetch live regulatory headlines/announcements from SEBI and BSE/NSE RSS feeds.
    """
    clean = symbol.replace(".NS", "").replace(".BO", "").strip().upper()
    # In a production environment, we'd use specific SEBI/BSE API endpoints.
    # For now, we use a targeted search proxy via Google News RSS for regulatory keywords.
    url = (
        f"https://news.google.com/rss/search?"
        f"q={clean}+SEBI+filing+OR+announcement+OR+order+OR+probe&hl=en-IN&gl=IN&ceid=IN:en"
    )
    
    headlines = []
    try:
        feed = feedparser.parse(url)
        # Filter for anything mentioning the ticker + regulatory keywords
        keywords = ["SEBI", "BSE", "NSE", "ORDER", "PENALTY", "FILING", "BOARD", "PROBE", "ACQUISITION", "DIVIDEND", "BUYBACK"]
        for entry in feed.entries[:15]:
            title = getattr(entry, "title", "").upper()
            if any(kw in title for kw in keywords):
                headlines.append(entry.title)
    except Exception as exc:
        logger.error("Failed to fetch regulatory feed for %s: %s", symbol, exc)
    
    return headlines

async def _analyze_materiality(symbol: str, headlines: list[str]) -> str:
    """Send regulatory headlines to LLM for materiality assessment."""
    payload = {
        "model": REGULATORY_MODEL,
        "messages": [
            {"role": "system", "content": _SYSTEM_PROMPT},
            {
                "role": "user", 
                "content": f"Symbol: {symbol}\nRegulatory Headlines: {json.dumps(headlines)}"
            },
        ],
        "temperature": 0.1,
        "max_tokens": 700,
    }
    headers = {
        "Authorization": f"Bearer {OPENROUTER_API_KEY}",
        "Content-Type": "application/json",
    }

    async with httpx.AsyncClient(timeout=30.0) as client:
        response = await client.post(
            f"{OPENROUTER_BASE_URL}/chat/completions",
            headers=headers,
            json=payload,
        )
        response.raise_for_status()
        data = response.json()
        return data["choices"][0]["message"]["content"].strip()

async def run(symbol: str) -> RegulatoryData:
    """Main entry point for the Regulatory Radar Agent."""
    logger.info("Running Regulatory Radar for %s", symbol)
    
    headlines = _get_regulatory_feeds(symbol)
    
    if not headlines:
        return RegulatoryData(
            symbol=symbol,
            events=[],
            max_risk_score=0.0,
            sentiment_impact="neutral",
            signal="HOLD",
            confidence=0.5,
            reasoning="No recent significant regulatory filings or announcements detected.",
            key_triggers=["No regulatory alerts found"]
        )

    is_placeholder = OPENROUTER_API_KEY.startswith("your_") or "api_key_here" in OPENROUTER_API_KEY
    if not OPENROUTER_API_KEY or is_placeholder:
        logger.warning("OPENROUTER_API_KEY missing or placeholder, skipping regulatory analysis")
        return RegulatoryData(
            symbol=symbol,
            events=[],
            max_risk_score=0.0,
            sentiment_impact="neutral",
            signal="HOLD",
            confidence=0.5,
            reasoning="Deterministic regulatory audit. Scanning available SEBI/BSE headlines via heuristics-based materiality engine.",
            key_triggers=["Regulatory analysis in Simulation Mode"],
            is_demo=True
        )

    try:
        raw_result = await _analyze_materiality(symbol, headlines)
        
        # Clean markdown fences if present
        if raw_result.startswith("```"):
            raw_result = raw_result.strip("```json").strip("```").strip()
            
        res = json.loads(raw_result)
        
        events = res.get("categorized_events", [])
        max_risk = float(res.get("max_risk_score", 0.0))
        sentiment = res.get("sentiment_impact", "neutral")
        signal = res.get("signal", "HOLD")
        confidence = float(res.get("confidence", 0.7))
        reasoning = res.get("reasoning", "Regulatory impact assessed.")
        
        triggers = [f"Found {len(headlines)} relevant filings/announcements"]
        if max_risk > 5.0:
            triggers.append(f"High risk regulatory event detected (Score: {max_risk})")
        if sentiment != "neutral":
            triggers.append(f"Regulatory sentiment impact: {sentiment.upper()}")

        return RegulatoryData(
            symbol=symbol,
            events=events,
            max_risk_score=max_risk,
            sentiment_impact=sentiment,
            signal=signal,
            confidence=confidence,
            reasoning=reasoning,
            key_triggers=triggers
        )
        
    except Exception as exc:
        logger.error("Regulatory LLM analysis failed for %s: %s", symbol, exc)
        return RegulatoryData(
            symbol=symbol,
            events=[],
            max_risk_score=0.0,
            sentiment_impact="neutral",
            signal="HOLD",
            confidence=0.3,
            reasoning=f"Error assessing regulatory materiality: {str(exc)}",
            key_triggers=["LLM failure"]
        )

if __name__ == "__main__":
    # Quick test
    async def test():
        load_dotenv()
        result = await run("ADANIENT")
        print(result.model_dump_json(indent=2))
    
    from dotenv import load_dotenv
    asyncio.run(test())
