backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
sse-starlette>=2.0.0
yfinance>=0.2.36
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
feedparser>=6.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
scipy>=1.10.0
